refactor(setup): extract the storage polyfill and reset it between tests
The polyfill had grown past what a setup file should hold inline, and its state
outlived every test in a file: a key written by one test was still there for the
next, so tests had to remember to clean up by hand (lazyNamed did, others were
simply lucky).

- move the polyfill to src/test/memoryStorage.ts; setup.ts now just calls
  installMemoryStorage()
- empty both areas after every test via resetMemoryStorage()
- reset the backing maps directly rather than calling clear(), so a test that
  leaves a throwing Storage.prototype spy behind cannot break the cleanup of the
  test after it
- drop the now-redundant hand cleanup in memoryStorage.test.ts

Written test-first: the isolation guard failed with "expected 'local' to be
null" before the reset existed, and fails again if the reset is removed. The
extraction itself changes no behavior and is covered by the existing contract
tests. Full suite: 295 files / 2092 tests, no test depended on storage
surviving between tests.
